This closes #3, reuse rune array instead of computing it every time (#2)
Converting a string to a rune array is not cheap. Since Golang strings
are immutable, each call to currentChar, doubleChar, etc. is doing
a lot of unnecessary work.

This commit fixes this by storing the conversion of string to runes
inside the parser and using it instead. This still incurs in the cost
of conversion, but it's done only once.